[front] cleanup: remove legacy agent observability UI
