--- front/components/agent_builder/AgentBuilderInsights.tsx
@@ -1,53 +1,0 @@
-import { useAgentBuilderContext } from "@app/components/agent_builder/AgentBuilderContext";
-import { AgentInsightsTab } from "@app/components/assistant/details/tabs/AgentInsightsTab";
-import { useAgentConfiguration } from "@app/lib/swr/assistants";
-import { cn, LoadingBlock } from "@dust-tt/sparkle";
-
-interface AgentBuilderInsightsProps {
-  agentConfigurationId: string;
-}
-
-export function AgentBuilderInsights({
-  agentConfigurationId,
-}: AgentBuilderInsightsProps) {
-  const { owner } = useAgentBuilderContext();
-
-  const { agentConfiguration, isAgentConfigurationLoading } =
-    useAgentConfiguration({
-      workspaceId: owner.sId,
-      agentConfigurationId,
-    });
-
-  if (isAgentConfigurationLoading || !agentConfiguration) {
-    return (
-      <div className="grid grid-cols-1 gap-6 p-4">
-        <ChartContainerSkeleton />
-        <ChartContainerSkeleton />
-        <ChartContainerSkeleton />
-      </div>
-    );
-  }
-
-  return (
-    <section className="flex h-full flex-col overflow-y-auto p-4">
-      <AgentInsightsTab owner={owner} agentConfiguration={agentConfiguration} />
-    </section>
-  );
-}
-
-function ChartContainerSkeleton() {
-  return (
-    <div
-      className={cn(
-        "bg-background flex flex-col rounded-lg border border-border p-4"
-      )}
-    >
-      <div className="mb-4 flex items-center justify-between">
-        <LoadingBlock className="h-6 w-40 rounded-md" />
-      </div>
-      <div className="flex-1">
-        <LoadingBlock className="h-full w-full rounded-xl" />
-      </div>
-    </div>
-  );
-}
