--- front/components/agent_builder/observability/charts/LatencyChart.tsx
@@ -1,261 +1,0 @@
-import {
-  LATENCY_LEGEND,
-  LATENCY_PALETTE,
-} from "@app/components/agent_builder/observability/constants";
-import type { LatencyPoint } from "@app/components/agent_builder/observability/hooks";
-import { useLatencyData } from "@app/components/agent_builder/observability/hooks";
-import { useObservabilityContext } from "@app/components/agent_builder/observability/ObservabilityContext";
-import { formatTimeSeriesTitle } from "@app/components/agent_builder/observability/shared/tooltipHelpers";
-import { VersionMarkersDots } from "@app/components/agent_builder/observability/shared/VersionMarkers";
-import { padSeriesToTimeRange } from "@app/components/agent_builder/observability/utils";
-import { ChartContainer } from "@app/components/charts/ChartContainer";
-import { legendFromConstant } from "@app/components/charts/ChartLegend";
-import { ChartTooltipCard } from "@app/components/charts/ChartTooltip";
-import { CHART_HEIGHT, CHART_MARGIN } from "@app/components/charts/constants";
-import { useSelectableSeries } from "@app/components/charts/useSelectableSeries";
-import type { AgentVersionMarker } from "@app/lib/api/assistant/observability/version_markers";
-import { useAgentVersionMarkers } from "@app/lib/swr/assistants";
-import { BROWSER_TIMEZONE } from "@app/lib/swr/workspaces";
-import { formatShortDate } from "@app/lib/utils/timestamps";
-import { cn } from "@dust-tt/sparkle";
-import { useMemo } from "react";
-import {
-  CartesianGrid,
-  Line,
-  LineChart,
-  Tooltip,
-  XAxis,
-  YAxis,
-} from "recharts";
-import type { TooltipContentProps } from "recharts/types/component/Tooltip";
-
-interface LatencyData extends LatencyPoint {
-  date: string;
-}
-
-function isLatencyData(data: unknown): data is LatencyData {
-  return (
-    typeof data === "object" &&
-    data !== null &&
-    "avgLatencyMs" in data &&
-    "percentilesLatencyMs" in data
-  );
-}
-
-function zeroFactory(timestamp: number) {
-  return {
-    timestamp,
-    count: 0,
-    avgLatencyMs: 0,
-    percentilesLatencyMs: 0,
-  };
-}
-
-function LatencyTooltip(
-  props: TooltipContentProps<number, string> & {
-    versionMarkers: AgentVersionMarker[];
-    activeKey?: string;
-    selectedKey?: string;
-  }
-) {
-  const { active, payload, versionMarkers, activeKey, selectedKey } = props;
-  if (!active || !payload || payload.length === 0) {
-    return null;
-  }
-  const first = payload[0];
-  if (!first?.payload || !isLatencyData(first.payload)) {
-    return null;
-  }
-  const row = first.payload;
-
-  return (
-    <ChartTooltipCard
-      title={formatTimeSeriesTitle(row.date, row.timestamp, versionMarkers)}
-      rows={[
-        {
-          key: "average",
-          label: "Average time",
-          value: `${row.avgLatencyMs}s`,
-          colorClassName: LATENCY_PALETTE.average,
-        },
-        {
-          key: "median",
-          label: "Median time",
-          value: `${row.percentilesLatencyMs}s`,
-          colorClassName: LATENCY_PALETTE.median,
-        },
-      ]}
-      activeKey={activeKey}
-      selectedKey={selectedKey}
-    />
-  );
-}
-
-interface LatencyChartProps {
-  workspaceId: string;
-  agentConfigurationId: string;
-  isCustomAgent: boolean;
-}
-
-export function LatencyChart({
-  workspaceId,
-  agentConfigurationId,
-  isCustomAgent,
-}: LatencyChartProps) {
-  const { period, mode, selectedVersion } = useObservabilityContext();
-
-  const {
-    data: rawData,
-    isLoading,
-    errorMessage,
-  } = useLatencyData({
-    workspaceId,
-    agentConfigurationId,
-    period,
-    mode: isCustomAgent ? mode : "timeRange",
-    filterVersion: isCustomAgent ? selectedVersion?.version : undefined,
-  });
-
-  const { versionMarkers } = useAgentVersionMarkers({
-    workspaceId,
-    agentConfigurationId,
-    days: period,
-    disabled: !workspaceId || !agentConfigurationId || !isCustomAgent,
-  });
-
-  const data = useMemo(() => {
-    if (mode === "timeRange") {
-      return padSeriesToTimeRange(
-        rawData,
-        mode,
-        period,
-        zeroFactory,
-        BROWSER_TIMEZONE
-      );
-    }
-
-    return rawData.map((data) => ({
-      ...data,
-      date: formatShortDate(data.timestamp),
-    }));
-  }, [rawData, mode, period]);
-
-  const {
-    selectedKey,
-    activeKey,
-    isDimmed,
-    lineActiveDot,
-    decorate,
-    hoverHandlers,
-  } = useSelectableSeries();
-
-  const legendItems = decorate(
-    legendFromConstant(LATENCY_LEGEND, LATENCY_PALETTE, {
-      includeVersionMarker:
-        isCustomAgent && mode === "timeRange" && versionMarkers.length > 0,
-    }),
-    { skip: (item) => item.key === "versionMarkers" }
-  );
-
-  return (
-    <ChartContainer
-      title="Latency"
-      description="Average and median time to complete output. Lower is better."
-      isLoading={isLoading}
-      errorMessage={errorMessage}
-      height={CHART_HEIGHT}
-      legendItems={legendItems}
-    >
-      <LineChart data={data} margin={CHART_MARGIN}>
-        <defs>
-          <linearGradient
-            id="fillAverage"
-            x1="0"
-            y1="0"
-            x2="0"
-            y2="1"
-            className={LATENCY_PALETTE.average}
-          >
-            <stop offset="5%" stopColor="currentColor" stopOpacity={0.8} />
-            <stop offset="95%" stopColor="currentColor" stopOpacity={0.1} />
-          </linearGradient>
-        </defs>
-        <CartesianGrid vertical={false} className="stroke-border" />
-        <XAxis
-          dataKey="date"
-          type="category"
-          scale="point"
-          allowDuplicatedCategory={false}
-          className="text-xs text-muted-foreground"
-          tickLine={false}
-          axisLine={false}
-          tickMargin={8}
-          minTickGap={16}
-        />
-        <YAxis
-          className="text-xs text-muted-foreground"
-          tickLine={false}
-          axisLine={false}
-          tickMargin={8}
-          tickFormatter={(value) => `${value}s`}
-          type="number"
-          allowDecimals={true}
-        />
-        <Tooltip
-          content={(props: TooltipContentProps<number, string>) => (
-            <LatencyTooltip
-              {...props}
-              versionMarkers={versionMarkers}
-              activeKey={activeKey}
-              selectedKey={selectedKey}
-            />
-          )}
-          cursor={false}
-          wrapperStyle={{ outline: "none", zIndex: 50 }}
-          contentStyle={{
-            background: "transparent",
-            border: "none",
-            padding: 0,
-            boxShadow: "none",
-          }}
-        />
-        <Line
-          type="monotone"
-          strokeWidth={2}
-          dataKey="avgLatencyMs"
-          name="Average time to complete output"
-          className={cn(
-            LATENCY_PALETTE.average,
-            "transition-opacity",
-            isDimmed("average") && "opacity-25"
-          )}
-          fill="url(#fillAverage)"
-          stroke="currentColor"
-          dot={false}
-          activeDot={lineActiveDot("average")}
-          isAnimationActive={false}
-          {...hoverHandlers("average")}
-        />
-        <Line
-          type="monotone"
-          strokeWidth={2}
-          dataKey="percentilesLatencyMs"
-          name="Median time to complete output"
-          className={cn(
-            LATENCY_PALETTE.median,
-            "transition-opacity",
-            isDimmed("median") && "opacity-25"
-          )}
-          stroke="currentColor"
-          dot={false}
-          activeDot={lineActiveDot("median")}
-          isAnimationActive={false}
-          {...hoverHandlers("median")}
-        />
-        {isCustomAgent && (
-          <VersionMarkersDots mode={mode} versionMarkers={versionMarkers} />
-        )}
-      </LineChart>
-    </ChartContainer>
-  );
-}
