--- front/components/agent_builder/observability/types.ts
@@ -1,92 +1,0 @@
-import type { AnalyticsVisibleOrigin } from "@app/components/agent_builder/observability/constants";
-
-const TOOL_CHART_MODES = ["version", "step"] as const;
-
-export type ToolChartModeType = (typeof TOOL_CHART_MODES)[number];
-
-const SKILL_CHART_MODES = ["version", "source"] as const;
-
-export type SkillChartModeType = (typeof SKILL_CHART_MODES)[number];
-
-export type ToolChartUsageDatum = {
-  percent: number;
-  count: number;
-  breakdown?: {
-    label: string;
-    count: number;
-    percent: number;
-  }[];
-};
-
-export type ChartDatum = {
-  label: string | number;
-  values: Record<string, ToolChartUsageDatum>;
-  total?: number;
-};
-
-type ToolChartUsagePayload = {
-  name?: string;
-  value?: number;
-  payload?: ChartDatum;
-};
-
-export type ToolLatencyDatum = {
-  name: string;
-  label: string;
-  count: number;
-  avgLatencyMs: number;
-  p50LatencyMs: number;
-  p95LatencyMs: number;
-};
-
-export type SourceChartDatum = {
-  origin: AnalyticsVisibleOrigin;
-  count: number;
-  percent: number;
-  label: string;
-};
-
-export type SkillSourceItem = {
-  skillName: string;
-  totalCount: number;
-  sources: Record<string, number>;
-};
-
-export function isSkillSourceItem(data: unknown): data is SkillSourceItem {
-  if (typeof data !== "object" || data === null) {
-    return false;
-  }
-  return (
-    "skillName" in data &&
-    typeof data.skillName === "string" &&
-    "totalCount" in data &&
-    typeof data.totalCount === "number" &&
-    "sources" in data &&
-    typeof data.sources === "object" &&
-    data.sources !== null
-  );
-}
-
-function isChartDatum(data: unknown): data is ChartDatum {
-  if (typeof data !== "object" || data === null) {
-    return false;
-  }
-  if (!("values" in data)) {
-    return false;
-  }
-  const v = (data as { values?: unknown }).values;
-  return !(typeof v !== "object" || v === null);
-}
-
-export function isToolChartUsagePayload(
-  d: unknown
-): d is ToolChartUsagePayload {
-  if (typeof d !== "object" || d === null) {
-    return false;
-  }
-  const o = d as Record<string, unknown>;
-  const valueOk = typeof o.value === "number";
-  const nameOk = typeof o.name === "string" || typeof o.name === "undefined";
-  const payloadOk = isChartDatum(o.payload);
-  return valueOk && nameOk && payloadOk;
-}
